Preprod

+++ frappe_s3_attachment/controller.py
@@ -15,7 +15,7 @@ import frappe
 
 
 import magic
-
+URL_PREFIXES = ("http://", "https://")
 
 class S3Operations(object):
 
@@ -97,37 +97,38 @@ class S3Operations(object):
         Uploads a new file to S3.
         Strips the file extension to set the content_type in metadata.
         """
-        mime_type = magic.from_file(file_path, mime=True)
-        key = self.key_generator(file_name, parent_doctype, parent_name, file_path)
-        content_type = mime_type
-        try:
-            if is_private:
-                self.S3_CLIENT.upload_file(
-                    file_path, self.BUCKET, key,
-                    ExtraArgs={
-                        "ContentType": content_type,
-                        "Metadata": {
+        if not file_path==None:
+            mime_type = magic.from_file(file_path, mime=True)
+            key = self.key_generator(file_name, parent_doctype, parent_name, file_path)
+            content_type = mime_type
+            try:
+                if is_private:
+                    self.S3_CLIENT.upload_file(
+                        file_path, self.BUCKET, key,
+                        ExtraArgs={
                             "ContentType": content_type,
-                            "file_name": file_name
+                            "Metadata": {
+                                "ContentType": content_type,
+                                "file_name": file_name
+                            }
                         }
-                    }
-                )
-            else:
-                self.S3_CLIENT.upload_file(
-                    file_path, self.BUCKET, key,
-                    ExtraArgs={
-                        "ContentType": content_type,
-                        # "ACL": 'public-read',
-                        "Metadata": {
+                    )
+                else:
+                    self.S3_CLIENT.upload_file(
+                        file_path, self.BUCKET, key,
+                        ExtraArgs={
                             "ContentType": content_type,
+                            # "ACL": 'public-read',
+                            "Metadata": {
+                                "ContentType": content_type,
 
+                            }
                         }
-                    }
-                )
+                    )
 
-        except boto3.exceptions.S3UploadFailedError:
-            frappe.throw(frappe._("File Upload Failed. Please try again."))
-        return key
+            except boto3.exceptions.S3UploadFailedError:
+                frappe.throw(frappe._("File Upload Failed. Please try again."))
+            return key
 
     def delete_from_s3(self, key):
         """Delete file from s3"""
@@ -203,35 +204,41 @@ def file_upload_to_s3(doc, method):
     ignore_s3_upload_for_doctype = frappe.local.conf.get('ignore_s3_upload_for_doctype') or ['Data Import']
     if parent_doctype not in ignore_s3_upload_for_doctype:
         if not doc.is_private:
-            file_path = site_path + '/public' + path
+            if  path.startswith("http"):
+                file_path =path
+            else:
+                file_path = site_path + '/public' + path
         else:
             file_path = site_path + path
-        key = s3_upload.upload_files_to_s3_with_key(
-            file_path, doc.file_name,
-            doc.is_private, parent_doctype,
-            parent_name
-        )
-
-        if doc.is_private:
-            method = "frappe_s3_attachment.controller.generate_file"
-            file_url = """/api/method/{0}?key={1}&file_name={2}""".format(method, key, doc.file_name)
-        else:
-            file_url = '{}/{}/{}'.format(
-                s3_upload.S3_CLIENT.meta.endpoint_url,
-                s3_upload.BUCKET,
-                key
+        # else:
+        #     file_path=doc.file_url
+        if not  path.startswith("http"):
+            key = s3_upload.upload_files_to_s3_with_key(
+                file_path, doc.file_name,
+                doc.is_private, parent_doctype,
+                parent_name
             )
-        os.remove(file_path)
-        frappe.db.sql("""UPDATE `tabFile` SET file_url=%s,
-            old_parent=%s, content_hash=%s WHERE name=%s""", (
-            file_url, 'Home/Attachments', key, doc.name))
-        
-        doc.file_url = file_url
-        
-        if parent_doctype and frappe.get_meta(parent_doctype).get('image_field'):
-            frappe.db.set_value(parent_doctype, parent_name, frappe.get_meta(parent_doctype).get('image_field'), file_url)
 
-        frappe.db.commit()
+            if doc.is_private:
+                method = "frappe_s3_attachment.controller.generate_file"
+                file_url = """/api/method/{0}?key={1}&file_name={2}""".format(method, key, doc.file_name)
+            else:
+                file_url = '{}/{}/{}'.format(
+                    s3_upload.S3_CLIENT.meta.endpoint_url,
+                    s3_upload.BUCKET,
+                    key
+                )
+            os.remove(file_path)
+            frappe.db.sql("""UPDATE `tabFile` SET file_url=%s,
+                old_parent=%s, content_hash=%s WHERE name=%s""", (
+                file_url, 'Home/Attachments', key, doc.name))
+            
+            doc.file_url = file_url
+            
+            if parent_doctype and frappe.get_meta(parent_doctype).get('image_field'):
+                frappe.db.set_value(parent_doctype, parent_name, frappe.get_meta(parent_doctype).get('image_field'), file_url)
+
+            frappe.db.commit()
 
 
 @frappe.whitelist()
